requirements.txt
playwright==1.40.0
pytest==7.4.3
pytest-playwright==0.4.3
allure-pytest==2.13.2
requests==2.31.0